+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * ZipUtil增加对输入流和字符串压缩的支持
 * Validator增加中国车牌号码
 * 新增StrUtil.cut方法，等分切割
+* 增加HttpUtil.download方法对https的支持
 
 ### Bug修复
 * 修复AbstractCache.onRemove回调方法在get和pruneCache时不被触发问题（issue#18@osc）